+++ src/modules/script-editor/host/browser-file-system.ts
@@ -4,6 +4,7 @@ import type {
   ScriptEditorSerializedTextFiles,
   ScriptEditorWriteResult,
 } from "./script-editor-file-system-host";
+export { createTextImportFilesFromRecord } from "./script-editor-file-system-host";
 
 export async function writeTextFilesWithDirectoryPicker(
   files: ScriptEditorSerializedTextFiles,
@@ -97,21 +98,6 @@ async function readFilesFromDirectoryHandleInternal(
   const files: File[] = [];
   await collectFilesFromDirectoryHandle(directoryHandle, "", files);
   return files;
-}
-
-export function createTextImportFilesFromRecord(
-  files: ScriptEditorSerializedTextFiles
-): File[] {
-  return Object.entries(files).map(([relativePath, content]) => {
-    const fileName = relativePath.split("/").filter(Boolean).pop() ?? relativePath;
-    const file = new File([content], fileName, {
-      type: "application/json",
-    });
-    Object.defineProperty(file, "webkitRelativePath", {
-      value: relativePath,
-    });
-    return file;
-  });
 }
 
 function getScriptEditorDirectoryPicker():